+++ src/test/java/org/sbrubbles/conditio/ScopeTest.java
@@ -2,7 +2,9 @@ package org.sbrubbles.conditio;
 
 import org.junit.jupiter.api.Test;
 import org.sbrubbles.conditio.fixtures.BasicCondition;
+import org.sbrubbles.conditio.fixtures.SonOfBasicCondition;
 import org.sbrubbles.conditio.policies.Policies;
+import org.sbrubbles.conditio.restarts.UseValue;
 
 import java.util.Collections;
 
@@ -40,8 +42,8 @@ public class ScopeTest {
     Scope root, first;
 
     try (Scope s1 = Scopes.create()) {
-        root = s1.getParent();
-        first = s1;
+      root = s1.getParent();
+      first = s1;
     }
 
     try (Scope s2 = Scopes.create()) {
@@ -54,7 +56,7 @@ public class ScopeTest {
   @Test
   public void createReflectsTheTryStack() {
     Scope root;
-    try(Scope s = Scopes.create()) {
+    try (Scope s = Scopes.create()) {
       root = s.getParent();
     }
 
@@ -80,10 +82,10 @@ public class ScopeTest {
 
   @Test
   public void createWithRestartsStoresThemOnlyDuringTheScopesLifetime() {
-    try(Scope a = Scopes.create()) {
+    try (Scope a = Scopes.create()) {
       assertIterableEquals(Collections.emptyList(), a.getAllRestarts());
 
-      try(Scope b = Scopes.create(Restarts.resume())) {
+      try (Scope b = Scopes.create(Restarts.resume())) {
         assertIterableEquals(Collections.emptyList(), a.getAllRestarts());
         assertIterableEquals(Collections.singletonList(Restarts.resume()), b.getAllRestarts());
       }
@@ -101,8 +103,34 @@ public class ScopeTest {
     }
   }
 
+  @Test
+  public void notifyHandlesHandlerNotFoundByResuming() {
+    try (Scope a = Scopes.create()) {
+      // signal throws...
+      assertThrows(HandlerNotFoundException.class, () -> a.signal(new BasicCondition(""), new Policies<>()));
+
+      // notify doesn't
+      a.notify(new BasicCondition(""));
+    }
+  }
+
+  @Test
+  public void notifyDoesntSwallowUnrelatedHandlerNotFounds() {
+    Condition c = new SonOfBasicCondition("");
+
+    try (Scope a = Scopes.create()) {
+      a.handle(SonOfBasicCondition.class, (s, ops) -> {
+        // a different, unhandled condition
+        Scope scope = s.getScope();
+        return ops.restart(new UseValue<>(scope.raise(new BasicCondition(""), String.class)));
+      });
+
+      assertThrows(HandlerNotFoundException.class, () -> a.notify(c));
+    }
+  }
+
   private Scope getClosedScope() {
-    try(Scope a = Scopes.create()) {
+    try (Scope a = Scopes.create()) {
       return a;
     }
   }